dom/tests/mochitest/pointerlock/file_retargetMouseEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_retargetMouseEvents.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>

  <div id="parent" style="width: 100%; height: 600px; background-color: blue;">
    <div id="child" style="width: 100%; height: 100%; background-color: red">
    </div>
  </div>

  <pre id="test">
    <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Retarget mouse events to the locked element
       */

      SimpleTest.waitForExplicitFinish();

      var parent, child;
      var pointer = navigator.mozPointer;

      /*
        The event listeners for the child element shouldn't be fired
        Mouse events will only happen when the pointer is locked
        and if the pointer is locked all the mouse events should be
        retargetted to the locked element
      */
      var childMoveTest = function() {
        ok(false, "Child shouldn't receive mousemove event");
      }

      var childDownTest = function() {
        ok(false, "Child shouldn't receive mousedown event");
      };

      var childUpTest = function() {
        ok(false, "Child shouldn't receive mouseup event");
      };

      var childClickTest = function() {
        ok(false, "Child shouldn't receive click event");
      };

      var childScrollTest = function() {
        // XXX: this is still not working...
        todo(false, "Child shouldn't receive scroll event");
      };

      /*
        Event listeners for the parent element
      */
      var parentMoveTest = function() {
        parent.removeEventListener("mousemove", parentMoveTest);
        child.removeEventListener("mousemove", childMoveTest);
        parent.addEventListener("DOMMouseScroll", parentScrollTest);
        child.addEventListener("DOMMouseScroll", childScrollTest);
        synthesizeMouseScroll(child, 5, 5, {'delta': 10, 'type': "DOMMouseScroll"});
      };

      var parentScrollTest = function (e) {
        ok(true, "Only the locked element should receive mousemove events");
        parent.removeEventListener("DOMMouseScroll", parentScrollTest);
        child.removeEventListener("DOMMouseScroll", childScrollTest);
        parent.addEventListener("mousedown", parentDownTest);
        child.addEventListener("mousedown", childDownTest);
        synthesizeMouseAtCenter(child, {type: "mousedown"}, window);
      };

      var parentDownTest = function (e) {
        ok(true, "Only the locked element should receive mousedown events");
        parent.removeEventListener("mousedown", parentDownTest);
        child.removeEventListener("mousedown", childDownTest);
        parent.addEventListener("mouseup", parentUpTest);
        child.addEventListener("mouseup", childUpTest);
        synthesizeMouseAtCenter(child, {type: "mouseup"}, window);
      };

      var parentUpTest = function (e) {
        ok(true, "Only the locked element should receive mouseup events");
        parent.removeEventListener("mouseup", parentUpTest);
        child.removeEventListener("mouseup", childUpTest);
        parent.addEventListener("click", parentClickTest);
        child.addEventListener("click", childClickTest);
        synthesizeMouseAtCenter(child, {type: "click"}, window);
      };

      var parentClickTest = function (e) {
        ok(true, "Only the locked element should receive click events");
        parent.removeEventListener("click", parentClickTest);
        child.removeEventListener("click", childClickTest);
        parent.addEventListener("mousemove", parentMoveTest);
        child.addEventListener("mousemove", childMoveTest);
        synthesizeMouseAtCenter(child, {type: "mousemove"}, window);
      };

      var parentMoveTest = function (e) {
        ok(true, "Only the locked element should receive scroll events");
        parent.removeEventListener("mousemove", parentMoveTest);
        child.removeEventListener("mousemove", childMoveTest);
        document.mozCancelFullScreen();
      }

      var successCallback = function () {
        parent.addEventListener("mousemove", parentMoveTest);
        child.addEventListener("mousemove", childMoveTest);
        synthesizeMouseAtCenter(parent, {type: "mousemove"}, window);
      };

      var failureCallback = function () {
        ok(false, "Failed to lock the pointer");
        document.mozCancelFullScreen();
      };

      document.addEventListener("mozfullscreenchange", function (e)  {
        if (document.mozFullScreen &&
            document.mozFullScreenElement === parent) {
          pointer.lock(parent, successCallback, failureCallback);
        } else {
          SimpleTest.finish();
        }
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          parent = document.getElementById("parent");
          child = document.getElementById("child");

          parent.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
